Edit Function
This is the last big thing that needs to be done for the backend part of the website. There is still an unknown issue of the blogs page that causes them to only re render when the page is reloaded. This is the next issue to fix as it greatly effects the usage of the website.